requirements.txt
streamlit
streamlit_gsheets
pandas
requests
Pillow